// src/graphql/generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type AuthenticateResponse = {
  __typename?: 'AuthenticateResponse';
  token: Scalars['String'];
};

export type Message = {
  __typename?: 'Message';
  id: Scalars['ID'];
  text: Scalars['String'];
  user: User;
};

export type MessageInput = {
  text?: Maybe<Scalars['String']>;
};

export type Mutation = {
  __typename?: 'Mutation';
  _empty?: Maybe<Scalars['String']>;
  createMessage: Message;
  deleteMessage: Scalars['Boolean'];
  register: AuthenticateResponse;
  updateMessage: UpdateMessageOutput;
};


export type MutationCreateMessageArgs = {
  text: Scalars['String'];
};


export type MutationDeleteMessageArgs = {
  id: Scalars['ID'];
};


export type MutationRegisterArgs = {
  email: Scalars['String'];
  password: Scalars['String'];
};


export type MutationUpdateMessageArgs = {
  id: Scalars['ID'];
  input: MessageInput;
};

export type Query = {
  __typename?: 'Query';
  _empty?: Maybe<Scalars['String']>;
  login: AuthenticateResponse;
  me: User;
  message: Message;
  messages: Array<Maybe<Message>>;
};


export type QueryLoginArgs = {
  email: Scalars['String'];
  password: Scalars['String'];
};


export type QueryMessageArgs = {
  id: Scalars['ID'];
};

export type Subscription = {
  __typename?: 'Subscription';
  _empty?: Maybe<Scalars['String']>;
};

export type UpdateMessageOutput = {
  __typename?: 'UpdateMessageOutput';
  success: Scalars['Boolean'];
  message?: Maybe<Message>;
};

export type User = {
  __typename?: 'User';
  id: Scalars['String'];
  username: Scalars['String'];
  messages?: Maybe<Array<Message>>;
};